requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements/dev.txt requirements/dev.in
#
alembic==1.4.2            # via flask-migrate
amqp==2.6.0               # via kombu
apache-superset==0.36.0   # via superset-patchup
apispec[yaml]==1.3.3      # via flask-appbuilder
appdirs==1.4.4            # via black, virtualenv
astroid==2.4.1            # via pylint, pylint-flask
attrs==19.3.0             # via black, jsonschema, pytest
babel==2.8.0              # via flask-babel
backcall==0.1.0           # via ipython
backoff==1.10.0           # via apache-superset
billiard==3.6.3.0         # via celery
black==19.10b0            # via -r requirements/dev.in
bleach==3.1.5             # via apache-superset
brotli==1.0.7             # via flask-compress
celery==4.4.4             # via apache-superset
cffi==1.14.0              # via cryptography
click==7.1.2              # via apache-superset, black, flask, flask-appbuilder, pip-tools
colorama==0.4.3           # via apache-superset, flask-appbuilder
contextlib2==0.6.0.post1  # via apache-superset
coverage==5.1             # via -r requirements/dev.in
croniter==0.3.32          # via apache-superset
cryptography==2.9.2       # via apache-superset
decorator==4.4.2          # via ipython, retry, traitlets
defusedxml==0.6.0         # via python3-openid
distlib==0.3.0            # via virtualenv
dnspython==1.16.0         # via email-validator
email-validator==1.1.1    # via flask-appbuilder
filelock==3.0.12          # via tox, virtualenv
flake8==3.8.2             # via -r requirements/dev.in
flask-appbuilder==2.3.4   # via apache-superset
flask-babel==1.0.0        # via flask-appbuilder
flask-caching==1.9.0      # via apache-superset
flask-compress==1.5.0     # via apache-superset
flask-jwt-extended==3.24.1  # via flask-appbuilder
flask-login==0.4.1        # via flask-appbuilder
flask-migrate==2.5.3      # via apache-superset
flask-openid==1.2.5       # via flask-appbuilder
flask-sqlalchemy==2.4.3   # via flask-appbuilder, flask-migrate
flask-talisman==0.7.0     # via apache-superset
flask-testing==0.8.0      # via -r requirements/dev.in
flask-wtf==0.14.3         # via apache-superset, flask-appbuilder
flask==1.1.2              # via apache-superset, flask-appbuilder, flask-babel, flask-caching, flask-compress, flask-jwt-extended, flask-login, flask-migrate, flask-openid, flask-sqlalchemy, flask-testing, flask-wtf
future==0.18.2            # via superset-patchup
geographiclib==1.50       # via geopy
geopy==1.22.0             # via apache-superset
gunicorn==20.0.4          # via apache-superset
humanize==2.4.0           # via apache-superset
idna==2.9                 # via email-validator
ipdb==0.13.2              # via -r requirements/dev.in
ipython-genutils==0.2.0   # via traitlets
ipython==7.15.0           # via ipdb
isodate==0.6.0            # via apache-superset
isort==4.3.21             # via -r requirements/dev.in, pylint
itsdangerous==1.1.0       # via flask, flask-wtf
jedi==0.17.0              # via ipython
jinja2==2.11.2            # via flask, flask-babel
jsonschema==3.2.0         # via flask-appbuilder
kombu==4.6.10             # via celery
lazy-object-proxy==1.4.3  # via astroid
mako==1.1.3               # via alembic
markdown==3.2.2           # via apache-superset
markupsafe==1.1.1         # via jinja2, mako, wtforms
marshmallow-enum==1.5.1   # via flask-appbuilder
marshmallow-sqlalchemy==0.23.1  # via flask-appbuilder
marshmallow==2.21.0       # via flask-appbuilder, marshmallow-enum, marshmallow-sqlalchemy
mccabe==0.6.1             # via flake8, pylint
more-itertools==8.3.0     # via pytest
msgpack==0.6.2            # via apache-superset
natsort==7.0.1            # via croniter
numpy==1.18.5             # via pandas, pyarrow
packaging==20.4           # via bleach, pytest, tox
pandas==0.25.3            # via apache-superset
parsedatetime==2.6        # via apache-superset
parso==0.7.0              # via jedi
pathlib2==2.3.5           # via apache-superset
pathspec==0.8.0           # via black
pep8==1.7.1               # via -r requirements/dev.in
pexpect==4.8.0            # via ipython
pickleshare==0.7.5        # via ipython
pip-tools==5.2.0          # via -r requirements/dev.in
pluggy==0.13.1            # via pytest, tox
polyline==1.4.0           # via apache-superset
prison==0.1.3             # via flask-appbuilder
prompt-toolkit==3.0.5     # via ipython
ptyprocess==0.6.0         # via pexpect
py==1.8.1                 # via pytest, retry, tox
pyarrow==0.16.0           # via apache-superset
pycodestyle==2.6.0        # via flake8
pycparser==2.20           # via cffi
pydocstyle==5.0.2         # via -r requirements/dev.in
pyflakes==2.2.0           # via flake8
pygments==2.6.1           # via ipython
pyjwt==1.7.1              # via flask-appbuilder, flask-jwt-extended
pylint-flask==0.6         # via -r requirements/dev.in
pylint-plugin-utils==0.6  # via pylint-flask
pylint==2.5.2             # via -r requirements/dev.in, pylint-flask, pylint-plugin-utils
pyparsing==2.4.7          # via packaging
pyrsistent==0.16.0        # via jsonschema
pytest==5.4.3             # via -r requirements/dev.in
python-dateutil==2.8.1    # via alembic, apache-superset, croniter, flask-appbuilder, pandas
python-dotenv==0.13.0     # via apache-superset
python-editor==1.0.4      # via alembic
python-geohash==0.8.5     # via apache-superset
python3-openid==3.1.0     # via flask-openid
pytz==2020.1              # via babel, celery, flask-babel, pandas
pyyaml==5.3.1             # via apache-superset, apispec
regex==2020.5.14          # via black
retry==0.9.2              # via apache-superset
selenium==3.141.0         # via apache-superset
simplejson==3.17.0        # via apache-superset
six==1.15.0               # via astroid, bleach, cryptography, flask-jwt-extended, flask-talisman, isodate, jsonschema, packaging, pathlib2, pip-tools, polyline, prison, pyarrow, pyrsistent, python-dateutil, sqlalchemy-utils, tox, traitlets, virtualenv, wtforms-json
snowballstemmer==2.0.0    # via pydocstyle
sqlalchemy-utils==0.36.6  # via apache-superset, flask-appbuilder
sqlalchemy==1.3.17        # via alembic, apache-superset, flask-sqlalchemy, marshmallow-sqlalchemy, sqlalchemy-utils
sqlparse==0.3.1           # via apache-superset
toml==0.10.1              # via black, pylint, tox
tox==3.15.1               # via -r requirements/dev.in
traitlets==4.3.3          # via ipython
typed-ast==1.4.1          # via black
urllib3==1.25.9           # via selenium
vine==1.3.0               # via amqp, celery
virtualenv==20.0.21       # via tox
wcwidth==0.2.3            # via prompt-toolkit, pytest
webencodings==0.5.1       # via bleach
werkzeug==1.0.1           # via flask, flask-jwt-extended
wrapt==1.12.1             # via astroid
wtforms-json==0.3.3       # via apache-superset
wtforms==2.3.1            # via flask-wtf, wtforms-json
yapf==0.30.0              # via -r requirements/dev.in

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
